LayoutTests/imported/w3c/web-platform-tests/dom/events/Event-defaultPrevented-after-dispatch-expected.txt


PASS Event.defaultPrevented is not reset after dipatchEvent() 
